requirements.txt
crewai==0.1.0
openai==1.3.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
reportlab==4.0.7
gspread==5.12.0
google-auth==2.23.4
